agentic-ai/1. Tasker - AI Sales Agent/code/ai-agent/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
azure-ai-projects
azure-identity
